提货扫描接口开发

+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/service/picking/JyPickingSendDestinationService.java
@@ -13,5 +13,5 @@ public interface JyPickingSendDestinationService {
      * @param nextSiteId
      * @return
      */
-    String fetchSendingBatchCode(Integer curSiteId, Integer nextSiteId);
+    String fetchSendingBatchCode(Integer curSiteId, Long nextSiteId);
 }
