+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/jy/service/picking/JyPickingSendDestinationServiceImpl.java
@@ -25,7 +25,7 @@ public class JyPickingSendDestinationServiceImpl implements JyPickingSendDestina
 
 
     @Override
-    public String fetchSendingBatchCode(Integer curSiteId, Integer nextSiteId) {
+    public String fetchSendingBatchCode(Integer curSiteId, Long nextSiteId) {
 
         //todo zcf
         return null;
